Ping for checking web server setting on tests is now scoped to components

+++ test/akelos/active_resource/cases/http_client.php
@@ -8,11 +8,13 @@ class HttpClient_TestCase extends ActiveResourceUnitTest
     public $verbs = array('get', 'post', 'put', 'delete');
     public $ht_access_path = '';
     public $original_ht_access = '';
+    public $webserver_enabled = false;
 
     function __construct()
     {
         if(!$this->webserver_enabled = AkConfig::getOption('webserver_enabled', false)){
-            echo "Skipping HttpClient_TestCase: Webserver no accesible at ".AkConfig::getOption('testing_url')."\n";
+            echo "Skipping HttpClient_TestCase: Web server not accesible at ".AkConfig::getOption('testing_url')."\n";
+            return;
         }
         $this->url = AkConfig::getOption('testing_url').
         '/active_resource/public/index.php?ak=http_requests';
@@ -26,6 +28,7 @@ class HttpClient_TestCase extends ActiveResourceUnitTest
 
     public function  test_get_verb()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual($this->Client->get($this->url), 'Hello unit tester');
         $this->assertEqual($this->Client->get($this->url.'/verb'), 'get');
         $this->assertEqual(Ak::url_get_contents($this->url.'/verb'), 'get');
@@ -33,36 +36,42 @@ class HttpClient_TestCase extends ActiveResourceUnitTest
 
     public function  test_post_verb()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual($this->Client->post($this->url.'/verb'), 'post');
         $this->assertEqual(Ak::url_get_contents($this->url.'/verb', array('method'=>'post')), 'post');
     }
 
     public function  test_put_verb()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual($this->Client->put($this->url.'/verb'), 'put');
         $this->assertEqual(Ak::url_get_contents($this->url.'/verb', array('method'=>'put')), 'put');
     }
 
     public function  test_delete_verb()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual($this->Client->delete($this->url.'/verb'), 'delete');
         $this->assertEqual(Ak::url_get_contents($this->url.'/verb', array('method'=>'delete')), 'delete');
     }
 
     public function test_should_get_response_header()
     {
+        if(!$this->webserver_enabled) return;
         $this->Client->get($this->url.'/test_header');
         $this->assertEqual($this->Client->getResponseHeader('x-test-header'), 'akelos');
     }
 
     public function test_should_get_response_code()
     {
+        if(!$this->webserver_enabled) return;
         $this->Client->get($this->url.'/code/201');
         $this->assertEqual($this->Client->getResponseCode(), 201);
     }
 
     public function test_should_set_user_agent()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual($this->Client->get($this->url.'/get_user_agent'), 'Akelos PHP Framework AkHttpClient (http://akelos.org)');
         $this->assertEqual(Ak::url_get_contents($this->url.'/get_user_agent'), 'Akelos PHP Framework AkHttpClient (http://akelos.org)');
         $this->assertEqual(Ak::url_get_contents($this->url.'/get_user_agent',
@@ -72,6 +81,7 @@ class HttpClient_TestCase extends ActiveResourceUnitTest
 
     public function test_should_send_params()
     {
+        if(!$this->webserver_enabled) return;
         $params = array('testing'=>array('user'=>'bermi','nested'=>array('one','two')));
         $expected = Ak::toJson($params['testing']);
 
@@ -85,11 +95,13 @@ class HttpClient_TestCase extends ActiveResourceUnitTest
 
     public function test_should_accept_redirects()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual(Ak::url_get_contents($this->url.'/redirect_1'), 3);
     }
 
     public function test_should_keep_cookies()
     {
+        if(!$this->webserver_enabled) return;
         $this->assertEqual(Ak::url_get_contents($this->url.'/persisting_cookies', array('cookies' => false)), 1);
         $this->assertEqual(Ak::url_get_contents($this->url.'/persisting_cookies', array('cookies' => true)), 1);
         $this->assertEqual(Ak::url_get_contents($this->url.'/persisting_cookies', array('cookies' => true)), 2);
